refactor(web): group contribution-service functions by concern
